requirements.txt
discord.py==2.4.0
requests
websockets
google-api-python-client
google-auth

